@@ -154,34 +154,34 @@
     "@types/uuid": "^8.3.4",
     "@urql/devtools": "^2.0.3",
     "@vue/composition-api": "^1.4.9",
-    "@vue/runtime-dom": "^3.2.31",
+    "@vue/runtime-dom": "^3.2.33",
     "@vue/test-utils": "^1.3.0",
-    "@vueuse/core": "^8.1.2",
+    "@vueuse/core": "^8.2.6",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^27.5.1",
-    "babel-loader": "^8.2.3",
+    "babel-loader": "^8.2.4",
     "chokidar": "^3.5.3",
-    "eslint": "^8.11.0",
+    "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-nuxt": "^3.2.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^8.5.0",
+    "eslint-plugin-vue": "^8.6.0",
     "jest": "^27.5.1",
     "jest-serializer-vue": "^2.0.2",
     "npm-run-all": "^4.1.5",
-    "nuxt-windicss": "^2.2.8",
-    "postcss-html": "^1.3.0",
-    "prettier": "^2.6.0",
+    "nuxt-windicss": "^2.2.11",
+    "postcss-html": "^1.4.1",
+    "prettier": "^2.6.2",
     "raw-loader": "^4.0.2",
-    "sass": "^1.49.9",
+    "sass": "^1.50.0",
     "sass-loader": "^10.2.0",
-    "stylelint": "^14.6.0",
+    "stylelint": "^14.7.1",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended-vue": "^1.3.0",
+    "stylelint-config-recommended-vue": "^1.4.0",
     "stylelint-config-standard": "^25.0.0",
     "stylelint-config-standard-scss": "^3.0.0",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.6.2",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.3",
     "vue-jest": "^3.0.7",
     "vue-template-babel-compiler": "^1.1.3",
     "worker-loader": "^3.0.8"
