@@ -1,6 +1,6 @@
 {
   "name": "@hoppscotch/js-sandbox",
-  "version": "2.0.0",
+  "version": "2.1.0",
   "description": "JavaScript sandboxes for running external scripts used by Hoppscotch clients",
   "main": "./lib/index.js",
   "module": "./lib/index.mjs",
@@ -40,27 +40,27 @@
   "author": "Hoppscotch (support@hoppscotch.io)",
   "license": "MIT",
   "dependencies": {
-    "@hoppscotch/data": "workspace:^0.4.0",
-    "fp-ts": "^2.11.9",
+    "@hoppscotch/data": "workspace:^0.4.2",
+    "fp-ts": "^2.11.10",
     "lodash": "^4.17.21",
     "quickjs-emscripten": "^0.15.0",
-    "tsup": "^5.11.13"
+    "tsup": "^5.12.5"
   },
   "devDependencies": {
     "@digitak/esrun": "^3.1.2",
-    "@relmify/jest-fp-ts": "^2.0.0",
+    "@relmify/jest-fp-ts": "^2.0.1",
     "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.180",
-    "@types/node": "^17.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.15.0",
-    "@typescript-eslint/parser": "^5.15.0",
-    "eslint": "^8.11.0",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.24",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
+    "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "io-ts": "^2.2.16",
     "jest": "^27.5.1",
-    "prettier": "^2.6.0",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.6.2"
+    "prettier": "^2.6.2",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.3"
   }
 }
