@@ -2,7 +2,7 @@
   <SmartModal
     v-if="show"
     dialog
-    :title="$t(`environment.${action}`)"
+    :title="t(`environment.${action}`)"
     @close="hideModal"
   >
     <template #body>
@@ -20,24 +20,24 @@
             @keyup.enter="saveEnvironment"
           />
           <label for="selectLabelEnvEdit">
-            {{ $t("action.label") }}
+            {{ t("action.label") }}
           </label>
         </div>
         <div class="flex items-center justify-between flex-1">
           <label for="variableList" class="p-4">
-            {{ $t("environment.variable_list") }}
+            {{ t("environment.variable_list") }}
           </label>
           <div class="flex">
             <ButtonSecondary
               v-tippy="{ theme: 'tooltip' }"
-              :title="$t('action.clear_all')"
+              :title="t('action.clear_all')"
               :svg="clearIcon"
               @click.native="clearContent()"
             />
             <ButtonSecondary
               v-tippy="{ theme: 'tooltip' }"
               svg="plus"
-              :title="$t('add.new')"
+              :title="t('add.new')"
               @click.native="addEnvironmentVariable"
             />
           </div>
@@ -46,7 +46,7 @@
           v-if="evnExpandError"
           class="w-full px-4 py-2 mb-2 overflow-auto font-mono text-red-400 whitespace-normal rounded bg-primaryLight"
         >
-          {{ $t("environment.nested_overflow") }}
+          {{ t("environment.nested_overflow") }}
         </div>
         <div class="border rounded divide-y divide-dividerLight border-divider">
           <div
@@ -57,12 +57,12 @@
             <input
               v-model="variable.key"
               class="flex flex-1 px-4 py-2 bg-transparent"
-              :placeholder="`${$t('count.variable', { count: index + 1 })}`"
+              :placeholder="`${t('count.variable', { count: index + 1 })}`"
               :name="'param' + index"
             />
             <SmartEnvInput
               v-model="variable.value"
-              :placeholder="`${$t('count.value', { count: index + 1 })}`"
+              :placeholder="`${t('count.value', { count: index + 1 })}`"
               :envs="liveEnvs"
               :name="'value' + index"
             />
@@ -70,7 +70,7 @@
               <ButtonSecondary
                 id="variable"
                 v-tippy="{ theme: 'tooltip' }"
-                :title="$t('action.remove')"
+                :title="t('action.remove')"
                 svg="trash"
                 color="red"
                 @click.native="removeEnvironmentVariable(index)"
@@ -85,13 +85,13 @@
               :src="`/images/states/${$colorMode.value}/blockchain.svg`"
               loading="lazy"
               class="inline-flex flex-col object-contain object-center w-16 h-16 my-4"
-              :alt="`${$t('empty.environments')}`"
+              :alt="`${t('empty.environments')}`"
             />
             <span class="pb-4 text-center">
-              {{ $t("empty.environments") }}
+              {{ t("empty.environments") }}
             </span>
             <ButtonSecondary
-              :label="`${$t('add.new')}`"
+              :label="`${t('add.new')}`"
               filled
               class="mb-4"
               @click.native="addEnvironmentVariable"
@@ -103,11 +103,11 @@
     <template #footer>
       <span>
         <ButtonPrimary
-          :label="`${$t('action.save')}`"
+          :label="`${t('action.save')}`"
           @click.native="saveEnvironment"
         />
         <ButtonSecondary
-          :label="`${$t('action.cancel')}`"
+          :label="`${t('action.cancel')}`"
           @click.native="hideModal"
         />
       </span>
@@ -115,9 +115,9 @@
   </SmartModal>
 </template>
 
-<script lang="ts">
+<script setup lang="ts">
 import clone from "lodash/clone"
-import { computed, defineComponent, PropType } from "@nuxtjs/composition-api"
+import { computed, ref, watch } from "@nuxtjs/composition-api"
 import * as E from "fp-ts/Either"
 import { Environment, parseTemplateStringE } from "@hoppscotch/data"
 import {
@@ -130,136 +130,147 @@ import {
   setGlobalEnvVariables,
   updateEnvironment,
 } from "~/newstore/environments"
-import { useReadonlyStream } from "~/helpers/utils/composables"
+import {
+  useReadonlyStream,
+  useI18n,
+  useToast,
+} from "~/helpers/utils/composables"
 
-export default defineComponent({
-  props: {
-    show: Boolean,
-    action: {
-      type: String as PropType<"new" | "edit">,
-      default: "edit",
-    },
-    editingEnvironmentIndex: {
-      type: [Number, String] as PropType<number | "Global" | null>,
-      default: null,
-    },
-    envVars: {
-      type: Function as PropType<() => Environment["variables"]>,
-      default: () => [],
-    },
-  },
-  setup(props) {
-    const globalVars = useReadonlyStream(globalEnv$, [])
+const t = useI18n()
+const toast = useToast()
 
-    const workingEnv = computed(() => {
-      if (props.editingEnvironmentIndex === "Global") {
-        return {
-          name: "Global",
-          variables: getGlobalVariables(),
-        } as Environment
-      } else if (props.action === "new") {
-        return {
-          name: "",
-          variables: props.envVars(),
-        }
-      } else if (props.editingEnvironmentIndex !== null) {
-        return getEnviroment(props.editingEnvironmentIndex)
-      } else {
-        return null
-      }
-    })
+const props = withDefaults(
+  defineProps<{
+    show: boolean
+    action: "edit" | "new"
+    editingEnvironmentIndex: number | "Global" | null
+    envVars: () => Environment["variables"]
+  }>(),
+  {
+    show: false,
+    action: "edit",
+    editingEnvironmentIndex: null,
+    envVars: () => [],
+  }
+)
 
+const emit = defineEmits<{
+  (e: "hide-modal"): void
+}>()
+
+const name = ref<string | null>(null)
+const vars = ref([{ key: "", value: "" }])
+const clearIcon = ref("trash-2")
+
+const globalVars = useReadonlyStream(globalEnv$, [])
+
+const workingEnv = computed(() => {
+  if (props.editingEnvironmentIndex === "Global") {
     return {
-      globalVars,
-      workingEnv,
-      envList: useReadonlyStream(environments$, []) || props.envVars(),
-    }
-  },
-  data() {
+      name: "Global",
+      variables: getGlobalVariables(),
+    } as Environment
+  } else if (props.action === "new") {
     return {
-      name: null as string | null,
-      vars: [] as { key: string; value: string }[],
-      clearIcon: "trash-2",
+      name: "",
+      variables: props.envVars(),
     }
-  },
-  computed: {
-    evnExpandError(): boolean {
-      for (const variable of this.vars) {
-        const result = parseTemplateStringE(variable.value, this.vars)
-
-        if (E.isLeft(result)) {
-          console.error("error", result.left)
-          return true
-        }
-      }
-      return false
-    },
-    liveEnvs(): Array<{ key: string; value: string; source: string }> {
-      if (this.evnExpandError) {
-        return []
-      }
-
-      if (this.$props.editingEnvironmentIndex === "Global") {
-        return [...this.vars.map((x) => ({ ...x, source: this.name! }))]
-      } else {
-        return [
-          ...this.vars.map((x) => ({ ...x, source: this.name! })),
-          ...this.globalVars.map((x) => ({ ...x, source: "Global" })),
-        ]
-      }
-    },
-  },
-  watch: {
-    show() {
-      this.name = this.workingEnv?.name ?? null
-      this.vars = clone(this.workingEnv?.variables ?? [])
-    },
-  },
-  methods: {
-    clearContent() {
-      this.vars = []
-      this.clearIcon = "check"
-      this.$toast.success(`${this.$t("state.cleared")}`)
-      setTimeout(() => (this.clearIcon = "trash-2"), 1000)
-    },
-    addEnvironmentVariable() {
-      this.vars.push({
-        key: "",
-        value: "",
-      })
-    },
-    removeEnvironmentVariable(index: number) {
-      this.vars.splice(index, 1)
-    },
-    saveEnvironment() {
-      if (!this.name) {
-        this.$toast.error(`${this.$t("environment.invalid_name")}`)
-        return
-      }
-
-      if (this.action === "new") {
-        createEnvironment(this.name)
-        setCurrentEnvironment(this.envList.length - 1)
-      }
-
-      const environmentUpdated: Environment = {
-        name: this.name,
-        variables: this.vars,
-      }
-
-      if (this.editingEnvironmentIndex === "Global")
-        setGlobalEnvVariables(environmentUpdated.variables)
-      else if (this.action === "new") {
-        updateEnvironment(this.envList.length - 1, environmentUpdated)
-      } else {
-        updateEnvironment(this.editingEnvironmentIndex!, environmentUpdated)
-      }
-      this.hideModal()
-    },
-    hideModal() {
-      this.name = null
-      this.$emit("hide-modal")
-    },
-  },
+  } else if (props.editingEnvironmentIndex !== null) {
+    return getEnviroment(props.editingEnvironmentIndex)
+  } else {
+    return null
+  }
 })
+
+const envList = useReadonlyStream(environments$, []) || props.envVars()
+
+const evnExpandError = computed(() => {
+  for (const variable of vars.value) {
+    const result = parseTemplateStringE(variable.value.toString(), vars.value)
+
+    if (E.isLeft(result)) {
+      console.error("error", result.left)
+      return true
+    }
+  }
+  return false
+})
+
+const liveEnvs = computed(() => {
+  if (evnExpandError) {
+    return []
+  }
+
+  if (props.editingEnvironmentIndex === "Global") {
+    return [...vars.value.map((x) => ({ ...x, source: name.value! }))]
+  } else {
+    return [
+      ...vars.value.map((x) => ({ ...x, source: name.value! })),
+      ...globalVars.value.map((x) => ({ ...x, source: "Global" })),
+    ]
+  }
+})
+
+watch(
+  () => props.show,
+  (show) => {
+    if (show) {
+      name.value = workingEnv.value?.name ?? null
+      vars.value = clone(workingEnv.value?.variables ?? [])
+    }
+  }
+)
+
+const clearContent = () => {
+  vars.value = []
+  clearIcon.value = "check"
+  toast.success(`${t("state.cleared")}`)
+  setTimeout(() => (clearIcon.value = "trash-2"), 1000)
+}
+
+const addEnvironmentVariable = () => {
+  vars.value.push({
+    key: "",
+    value: "",
+  })
+}
+
+const removeEnvironmentVariable = (index: number) => {
+  vars.value.splice(index, 1)
+}
+
+const saveEnvironment = () => {
+  if (!name.value) {
+    toast.error(`${t("environment.invalid_name")}`)
+    return
+  }
+
+  const environmentUpdated: Environment = {
+    name: name.value,
+    variables: vars.value,
+  }
+
+  if (props.action === "new") {
+    // Creating a new environment
+    createEnvironment(name.value)
+    updateEnvironment(envList.value.length - 1, environmentUpdated)
+    setCurrentEnvironment(envList.value.length - 1)
+    toast.success(`${t("environment.created")}`)
+  } else if (props.editingEnvironmentIndex === "Global") {
+    // Editing the Global environment
+    setGlobalEnvVariables(environmentUpdated.variables)
+    toast.success(`${t("environment.updated")}`)
+  } else if (props.editingEnvironmentIndex !== null) {
+    // Editing an environment
+    updateEnvironment(props.editingEnvironmentIndex, environmentUpdated)
+    toast.success(`${t("environment.updated")}`)
+  }
+
+  hideModal()
+}
+
+const hideModal = () => {
+  name.value = null
+  emit("hide-modal")
+}
 </script>
